Handle old setuptools namespace package protocol
